requirements.txt
Flask
flask_sqlalchemy
flask_login
flask_wtf
WTForms
wtforms
email_validator
bcrypt
cryptography
python-dotenv
pyotp